"move selector methods to plot"
