+++ bokeh/plotting/helpers.py
@@ -321,17 +321,6 @@ def _process_tools_arg(plot, tools):
 
     return tool_objs
 
-class _list_attr_splat(list):
-    def __setattr__(self, attr, value):
-        for x in self:
-            setattr(x, attr, value)
-
-    def __dir__(self):
-        if len(set(type(x) for x in self)) == 1:
-            return dir(self[0])
-        else:
-            return dir(self)
-
 _arg_template = "    %s (%s) : %s (default %r)"
 _doc_template = """ Configure and add %s glyphs to this Figure.
